+++ .github/workflows/unittest.yml
@@ -72,6 +72,30 @@ jobs:
           parallel: true
           flag-name: run ${{ env.CURRENT_NODE_VERSION }}
           fail-on-error: false
+  framework-integration:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Use Node.js ${{ env.CURRENT_NODE_VERSION }}
+        uses: actions/setup-node@v4.0.1
+        with:
+          node-version: ${{ env.CURRENT_NODE_VERSION }}
+          cache: 'npm'
+          cache-dependency-path: |
+            framework/package-lock.json
+            website/package-lock.json
+            package-lock.json
+      - run: cd framework && npm ci
+      - run: cd framework && npm run-script integration
+      - run: mv ./framework/coverage ./
+      - run: sed -i 's/SF:src/SF:framework\/src/' ./coverage/lcov.info
+      - name: Coveralls
+        uses: coverallsapp/github-action@master
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          parallel: true
+          flag-name: run ${{ env.CURRENT_NODE_VERSION }}
+          fail-on-error: false
   website:
     runs-on: ubuntu-latest
     steps:
@@ -125,6 +149,7 @@ jobs:
     needs:
     - framework-previous
     - framework-current
+    - framework-integration
     - website
     - e2e
     steps:
