+++ framework/package.json
@@ -86,7 +86,8 @@
     "lint": "eslint . --ext .js,.ts,.cjs,.json,.tsx",
     "lint-fix": "eslint --fix . --ext .js,.ts,.cjs,.json",
     "lint-md": "markdownlint *.md",
-    "test": "c8 mocha",
+    "test": "c8 mocha test",
+    "integration": "c8 mocha integration",
     "clear": "node tools/clear-js-files-from-src.js",
     "tsc": "tsc -p tsconfig.json",
     "prepublishOnly": "npm run tsc",
